Squash all changes on top of main (#3)
* Squash all changes on top of main

* Even higher tolerance for Spacing error in images

* Fix issue with sitk parameters

* Lower tolerance for Image parameters checking

* Fix saveSteps bug

* Increase tolerance for test_Siddon

* Fix bug with MLEM Warper

+++ yrt-pet/include/operators/TimeOfFlight.hpp
@@ -16,9 +16,9 @@ class TimeOfFlightHelper
 public:
 	explicit TimeOfFlightHelper(float tof_width_ps, int tof_n_std = -1);
 
-	HOST_DEVICE_CALLABLE inline void getAlphaRange(double& alpha_min,
-	                                               double& alpha_max,
-	                                               double lorNorm,
+	HOST_DEVICE_CALLABLE inline void getAlphaRange(float& alpha_min,
+	                                               float& alpha_max,
+	                                               float lorNorm,
 	                                               float tofValue_ps) const
 	{
 		using namespace std;
@@ -29,7 +29,7 @@ public:
 		}
 		else
 		{
-			const double tof_value_mm = tofValue_ps * SPEED_OF_LIGHT_MM_PS * 0.5;
+			const float tof_value_mm = tofValue_ps * SPEED_OF_LIGHT_MM_PS * 0.5;
 			alpha_min =
 			    max(0.0, 0.5 + (tof_value_mm - m_truncWidth_mm) / lorNorm);
 			alpha_max =
@@ -37,13 +37,13 @@ public:
 		}
 	}
 
-	HOST_DEVICE_CALLABLE inline float getWeight(double lorNorm,
+	HOST_DEVICE_CALLABLE inline float getWeight(float lorNorm,
 	                                            float tofValue_ps,
 	                                            float offLo_mm,
 	                                            float offHi_mm) const
 	{
 		const float tof_value_mm = tofValue_ps * SPEED_OF_LIGHT_MM_PS * 0.5;
-		const double pc = 0.5 * lorNorm + tof_value_mm;
+		const float pc = 0.5 * lorNorm + tof_value_mm;
 
 		const float x_cent_norm = (0.5f * (offLo_mm + offHi_mm) - pc) / m_sigma;
 		return exp(-0.5f * x_cent_norm * x_cent_norm) * m_norm;
